requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cleanup==8.1.0
django-decouple==2.1
django-router==1.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
install==1.3.5
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
